+++ src/main/java/cart/web/service/CartService.java
@@ -1,5 +1,7 @@
-package cart.domain.cart;
+package cart.web.service;
 
+import cart.domain.cart.CartProduct;
+import cart.domain.cart.CartRepository;
 import cart.domain.user.User;
 import cart.domain.user.UserRepository;
 import cart.exception.UserNotFoundException;
