+++ src/main/java/cart/web/service/UserService.java
@@ -1,5 +1,7 @@
-package cart.domain.user;
+package cart.web.service;
 
+import cart.domain.user.User;
+import cart.domain.user.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
