+++ src/test/java/cart/web/controller/product/ProductViewControllerTest.java
@@ -2,8 +2,8 @@ package cart.web.controller.product;
 
 import cart.domain.product.Product;
 import cart.domain.product.ProductCategory;
-import cart.domain.product.ProductService;
 import cart.web.controller.admin.AdminViewController;
+import cart.web.service.ProductService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
